+++ packages/core/stellar/src/zip-rust.ts
@@ -4,13 +4,13 @@ import type { Contract } from './contract';
 import { printContract, removeCreateLevelAttributes } from './print';
 import {
   addDependenciesWith,
-  allStellarDependencies,
+  libDependencies,
   contractOptionsToContractName,
   createRustLibFile,
   printContractCargo,
   printRustNameTest,
 } from './zip-shared';
-import { contractsVersionTag, compatibleSorobanVersion } from './utils/version';
+import { contractsVersion, compatibleSorobanVersion } from './utils/version';
 
 const workspaceCargo = `[workspace]
 resolver = "2"
@@ -23,7 +23,7 @@ license = "Apache-2.0"
 version = "0.0.1"
 
 [workspace.dependencies]
-${addDependenciesWith(`{ git = "https://github.com/OpenZeppelin/stellar-contracts", tag = "${contractsVersionTag}" }`, allStellarDependencies)}${addDependenciesWith(`{ version = "${compatibleSorobanVersion}" }`, ['soroban'])}
+${addDependenciesWith(`"${compatibleSorobanVersion}"`, ['soroban-sdk'])}${addDependenciesWith(`"=${contractsVersion}"`, [...libDependencies])}
 `;
 
 const readme = `\
@@ -41,7 +41,7 @@ Continue your development journey with [Stellar CLI](https://github.com/stellar/
 - See [Git installation guide](https://github.com/git-guides/install-git).
 `;
 
-export async function zipRust(c: Contract, opts: GenericOptions) {
+export const createRustZipEnvironment = (c: Contract, opts: GenericOptions) => {
   const zip = new JSZip();
 
   const contractName = contractOptionsToContractName(opts?.kind || 'contract');
@@ -51,7 +51,11 @@ export async function zipRust(c: Contract, opts: GenericOptions) {
   zip.file(`contracts/${contractName}/src/lib.rs`, createRustLibFile);
   zip.file(`contracts/${contractName}/Cargo.toml`, printContractCargo(contractName));
   zip.file('Cargo.toml', workspaceCargo);
-  zip.file('README.md', readme);
 
   return zip;
-}
+};
+
+const addRustProjectReadme = (zip: JSZip) => zip.file('README.md', readme);
+
+export const zipRustProject = async (c: Contract, opts: GenericOptions) =>
+  addRustProjectReadme(createRustZipEnvironment(c, opts));