+++ packages/core/stellar/src/zip-shared.ts
@@ -1,4 +1,5 @@
 import type { Contract } from './contract';
+import { contractsVersion, compatibleSorobanVersion } from './utils/version';
 
 function pascalToSnakeCase(string: string) {
   return string
@@ -38,28 +39,17 @@ fn initial_state() {
 // Add more tests bellow
 `;
 
-const getKeysOf = <TObject extends Record<string, unknown>>(objectToGetKeysOf: TObject) =>
-  Object.keys(objectToGetKeysOf) as (keyof TObject)[];
+export const libDependencies = [
+  'stellar-tokens',
+  'stellar-access',
+  'stellar-contract-utils',
+  'stellar-macros',
+] as const;
 
-export const stellarDependencies = {
-  base: ['stellar-default-impl-macro'],
-  fungible: ['stellar-fungible'],
-  nonFungible: ['stellar-non-fungible'],
-  pausable: ['stellar-pausable', 'stellar-pausable-macros'],
-  upgradable: ['stellar-upgradeable', 'stellar-upgradeable-macros'],
-} as const;
+export const allDependencies = [...libDependencies, 'soroban-sdk'] as const;
 
-export const allStellarDependencies = getKeysOf(stellarDependencies);
-
-export const allDependencies = {
-  ...stellarDependencies,
-  soroban: ['soroban-sdk'],
-} as const;
-
-export const addDependenciesWith = (dependencyValue: string, dependenciesToAdd: (keyof typeof allDependencies)[]) =>
-  dependenciesToAdd.reduce((addedDependency, dependencyName) => {
-    return `${addedDependency}${allDependencies[dependencyName].map(cargoDependencies => `${cargoDependencies} = ${dependencyValue}\n`).join('')}`;
-  }, '');
+export const addDependenciesWith = (dependencyValue: string, dependenciesToAdd: string[]) =>
+  `${dependenciesToAdd.map(dependency => `${dependency} = ${dependencyValue}\n`).join('')}`;
 
 export const printContractCargo = (scaffoldContractName: string) => `[package]
 name = "${scaffoldContractName.replace(/_/, '-')}-contract"
@@ -73,13 +63,27 @@ crate-type = ["cdylib"]
 doctest = false
 
 [dependencies]
-${addDependenciesWith('{ workspace = true }', [...allStellarDependencies, 'soroban'])}
+${addDependenciesWith('{ workspace = true }', [...allDependencies])}
 [dev-dependencies]
-${addDependenciesWith('{ workspace = true, features = ["testutils"] }', ['soroban'])}`;
+${addDependenciesWith('{ workspace = true, features = ["testutils"] }', ['soroban-sdk'])}`;
 
 export const createRustLibFile = `#![no_std]
 #![allow(dead_code)]
 
 mod contract;
 mod test;
+`;
+
+export const workspaceCargo = `[workspace]
+resolver = "2"
+members = ["contracts/*"]
+
+[workspace.package]
+authors = []
+edition = "2021"
+license = "Apache-2.0"
+version = "0.0.1"
+
+[workspace.dependencies]
+${addDependenciesWith(`"${compatibleSorobanVersion}"`, ['soroban-sdk'])}${addDependenciesWith(`"=${contractsVersion}"`, [...libDependencies])}
 `;