+++ packages/core/stylus/src/erc20.ts
@@ -1,17 +1,16 @@
-import type { BaseImplementedTrait, Contract } from './contract';
+import type { Contract, ContractTrait, StoredContractTrait } from './contract';
 import { ContractBuilder } from './contract';
-import { defineFunctions } from './utils/define-functions';
 import type { CommonContractOptions } from './common-options';
 import { withCommonContractDefaults, getSelfArg } from './common-options';
 import { contractDefaults as commonDefaults } from './common-options';
 import { printContract } from './print';
 import { setAccessControl } from './set-access-control';
 import { setInfo } from './set-info';
+import { defineFunctions } from './utils/define-functions';
 
 export interface ERC20Options extends CommonContractOptions {
   name: string;
   burnable?: boolean;
-  // pausable?: boolean;
   permit?: boolean;
   flashmint?: boolean;
 }
@@ -19,7 +18,6 @@ export interface ERC20Options extends CommonContractOptions {
 export const defaults: Required<ERC20Options> = {
   name: 'MyToken',
   burnable: false,
-  // pausable: false,
   permit: true,
   flashmint: false,
   access: commonDefaults.access,
@@ -35,7 +33,6 @@ function withDefaults(opts: ERC20Options): Required<ERC20Options> {
     ...opts,
     ...withCommonContractDefaults(opts),
     burnable: opts.burnable ?? defaults.burnable,
-    // pausable: opts.pausable ?? defaults.pausable,
     permit: opts.permit ?? defaults.permit,
     flashmint: opts.flashmint ?? defaults.flashmint,
   };
@@ -52,10 +49,6 @@ export function buildERC20(opts: ERC20Options): Contract {
   const allOpts = withDefaults(opts);
 
   addBase(c);
-
-  // if (allOpts.pausable) {
-  //   addPausable(c, allOpts.access);
-  // }
 
   if (allOpts.burnable) {
     addBurnable(c);
@@ -77,188 +70,213 @@ export function buildERC20(opts: ERC20Options): Contract {
 
 function addBase(c: ContractBuilder) {
   c.addImplementedTrait(erc20Trait);
-  // c.addImplementedTrait(erc20MetadataTrait);
-
-  // the trait necessary to access Erc20 functions within custom functions of the child contract
-  c.addUseClause('openzeppelin_stylus::token::erc20', 'IErc20');
-
-  // if (pausable) {
-  //   c.addUseClause('alloc::vec', 'Vec');
-  //   c.addUseClause('alloy_primitives', 'Address');
-  //   c.addUseClause('alloy_primitives', 'U256');
-
-  //   c.addFunctionCodeBefore(erc20Trait, functions.transfer, ['self.pausable.when_not_paused()?;']);
-  //   c.addFunctionCodeBefore(erc20Trait, functions.transfer_from, ['self.pausable.when_not_paused()?;']);
-  // }
 }
 
 function addPermit(c: ContractBuilder) {
-  c.addImplementedTrait(erc20PermitTrait);
   c.addImplementedTrait(noncesTrait);
+  c.addImplementedTrait(permitTrait);
   c.addEip712();
-
-  c.addUseClause('alloc::vec', 'Vec');
-  c.addUseClause('alloy_primitives', 'Address');
-  c.addUseClause('alloy_primitives', 'B256');
-  c.addUseClause('alloy_primitives', 'U256');
-
-  c.addFunction(erc20PermitTrait, functions.permit);
-
-  // if (pausable) {
-  //   c.addFunctionCodeBefore(erc20PermitTrait, functions.permit, ['self.pausable.when_not_paused()?;']);
-  // }
 }
 
 function addBurnable(c: ContractBuilder) {
-  c.addUseClause('openzeppelin_stylus::token::erc20::extensions', 'IErc20Burnable');
-
-  c.addUseClause('alloc::vec', 'Vec');
-  c.addUseClause('alloy_primitives', 'Address');
-  c.addUseClause('alloy_primitives', 'U256');
-
-  c.addFunction(erc20Trait, functions.burn);
-  c.addFunction(erc20Trait, functions.burn_from);
-
-  // if (pausable) {
-  //   c.addFunctionCodeBefore(erc20Trait, functions.burn, ['self.pausable.when_not_paused()?;']);
-  //   c.addFunctionCodeBefore(erc20Trait, functions.burn_from, ['self.pausable.when_not_paused()?;']);
-  // }
+  c.addImplementedTrait(burnableTrait);
 }
 
 function addFlashMint(c: ContractBuilder) {
   c.addImplementedTrait(flashMintTrait);
-
-  // the trait necessary to access Erc20FlashMint functions within custom functions of the child contract
-  c.addUseClause('openzeppelin_stylus::token::erc20::extensions', 'IErc3156FlashLender');
-
-  c.addUseClause('stylus_sdk', 'abi::Bytes');
-  c.addUseClause('alloy_primitives', 'Address');
-  c.addUseClause('alloy_primitives', 'U256');
-
-  c.addFunction(flashMintTrait, functions.max_flash_loan);
-  c.addFunction(flashMintTrait, functions.flash_fee);
-  c.addFunction(flashMintTrait, functions.flash_loan);
-
-  // if (pausable) {
-  //   c.addFunctionCodeBefore(flashMintTrait, functions.flash_loan, ['self.pausable.when_not_paused()?;']);
-  // }
 }
 
-const erc20Trait: BaseImplementedTrait = {
-  name: 'Erc20',
+const ERC20_STORAGE_NAME = 'erc20';
+const NONCES_STORAGE_NAME = 'nonces';
+const PERMIT_STORAGE_NAME = 'erc20_permit';
+const FLASH_MINT_STORAGE_NAME = 'flash_mint';
+
+const functions = {
+  erc20: defineFunctions({
+    total_supply: {
+      args: [getSelfArg('immutable')],
+      returns: 'U256',
+      code: `self.${ERC20_STORAGE_NAME}.total_supply()`,
+    },
+    balance_of: {
+      args: [getSelfArg('immutable'), { name: 'account', type: 'Address' }],
+      returns: 'U256',
+      code: `self.${ERC20_STORAGE_NAME}.balance_of(account)`,
+    },
+    transfer: {
+      args: [getSelfArg(), { name: 'to', type: 'Address' }, { name: 'value', type: 'U256' }],
+      returns: { ok: 'bool', err: 'Self::Error' },
+      code: `self.${ERC20_STORAGE_NAME}.transfer(to, value)?`,
+    },
+    allowance: {
+      args: [getSelfArg('immutable'), { name: 'owner', type: 'Address' }, { name: 'spender', type: 'Address' }],
+      returns: 'U256',
+      code: `self.${ERC20_STORAGE_NAME}.allowance(owner, spender)`,
+    },
+    approve: {
+      args: [getSelfArg(), { name: 'spender', type: 'Address' }, { name: 'value', type: 'U256' }],
+      returns: { ok: 'bool', err: 'Self::Error' },
+      code: `self.${ERC20_STORAGE_NAME}.approve(spender, value)?`,
+    },
+    transfer_from: {
+      args: [
+        getSelfArg(),
+        { name: 'from', type: 'Address' },
+        { name: 'to', type: 'Address' },
+        { name: 'value', type: 'U256' },
+      ],
+      returns: { ok: 'bool', err: 'Self::Error' },
+      code: `self.${ERC20_STORAGE_NAME}.transfer_from(from, to, value)?`,
+    },
+  }),
+  nonces: defineFunctions({
+    nonces: {
+      args: [getSelfArg('immutable'), { name: 'owner', type: 'Address' }],
+      returns: 'U256',
+      code: `self.${NONCES_STORAGE_NAME}.nonces(owner)`,
+    },
+  }),
+  permit: defineFunctions({
+    domain_separator: {
+      attribute: 'selector(name = "DOMAIN_SEPARATOR")',
+      args: [getSelfArg('immutable')],
+      returns: 'B256',
+      code: `self.${PERMIT_STORAGE_NAME}.domain_separator()`,
+    },
+    permit: {
+      args: [
+        getSelfArg(),
+        { name: 'owner', type: 'Address' },
+        { name: 'spender', type: 'Address' },
+        { name: 'value', type: 'U256' },
+        { name: 'deadline', type: 'U256' },
+        { name: 'v', type: 'u8' },
+        { name: 'r', type: 'B256' },
+        { name: 's', type: 'B256' },
+      ],
+      returns: { ok: '()', err: 'Self::Error' },
+      code: `self.${PERMIT_STORAGE_NAME}.permit(owner, spender, value, deadline, v, r, s, &mut self.${ERC20_STORAGE_NAME}, &mut self.${NONCES_STORAGE_NAME})?`,
+    },
+  }),
+  burnable: defineFunctions({
+    burn: {
+      args: [getSelfArg(), { name: 'value', type: 'U256' }],
+      returns: { ok: '()', err: 'Self::Error' },
+      code: `self.${ERC20_STORAGE_NAME}.burn(value)?`,
+    },
+    burn_from: {
+      args: [getSelfArg(), { name: 'account', type: 'Address' }, { name: 'value', type: 'U256' }],
+      returns: { ok: '()', err: 'Self::Error' },
+      code: `self.${ERC20_STORAGE_NAME}.burn_from(account, value)?`,
+    },
+  }),
+  flash_mint: defineFunctions({
+    max_flash_loan: {
+      args: [getSelfArg('immutable'), { name: 'token', type: 'Address' }],
+      returns: 'U256',
+      code: `self.${FLASH_MINT_STORAGE_NAME}.max_flash_loan(token, &self.${ERC20_STORAGE_NAME})`,
+    },
+    flash_fee: {
+      args: [getSelfArg('immutable'), { name: 'token', type: 'Address' }, { name: 'value', type: 'U256' }],
+      returns: { ok: 'U256', err: 'Self::Error' },
+      code: `self.${FLASH_MINT_STORAGE_NAME}.flash_fee(token, value)?`,
+    },
+    flash_loan: {
+      args: [
+        getSelfArg(),
+        { name: 'receiver', type: 'Address' },
+        { name: 'token', type: 'Address' },
+        { name: 'value', type: 'U256' },
+        { name: 'data', type: 'Bytes' },
+      ],
+      returns: { ok: 'bool', err: 'Self::Error' },
+      code: `self.${FLASH_MINT_STORAGE_NAME}.flash_loan(receiver, token, value, &data, &mut self.${ERC20_STORAGE_NAME})?`,
+    },
+  }),
+};
+
+const erc20Trait: StoredContractTrait = {
+  name: 'IErc20',
+  associatedError: true,
+  errors: [
+    { variant: 'InsufficientBalance', value: { module: 'erc20', error: 'ERC20InsufficientBalance' } },
+    { variant: 'InvalidSender', value: { module: 'erc20', error: 'ERC20InvalidSender' } },
+    { variant: 'InvalidReceiver', value: { module: 'erc20', error: 'ERC20InvalidReceiver' } },
+    { variant: 'InsufficientAllowance', value: { module: 'erc20', error: 'ERC20InsufficientAllowance' } },
+    { variant: 'InvalidSpender', value: { module: 'erc20', error: 'ERC20InvalidSpender' } },
+    { variant: 'InvalidApprover', value: { module: 'erc20', error: 'ERC20InvalidApprover' } },
+  ],
   storage: {
-    name: 'erc20',
+    name: ERC20_STORAGE_NAME,
     type: 'Erc20',
   },
   modulePath: 'openzeppelin_stylus::token::erc20',
+  requiredImports: [
+    { containerPath: 'alloc::vec', name: 'Vec' },
+    { containerPath: 'stylus_sdk::alloy_primitives', name: 'Address' },
+    { containerPath: 'stylus_sdk::alloy_primitives', name: 'U256' },
+  ],
+  functions: Object.values(functions.erc20),
 };
 
-const erc20PermitTrait: BaseImplementedTrait = {
-  name: 'Erc20Permit',
+const noncesTrait: StoredContractTrait = {
+  name: 'INonces',
   storage: {
-    name: 'erc20_permit',
-    type: 'Erc20Permit<Eip712>',
-  },
-  modulePath: 'openzeppelin_stylus::token::erc20::extensions',
-};
-
-const flashMintTrait: BaseImplementedTrait = {
-  name: 'Erc20FlashMint',
-  storage: {
-    name: 'flash_mint',
-    type: 'Erc20FlashMint',
-  },
-  omitInherit: true,
-  modulePath: 'openzeppelin_stylus::token::erc20::extensions',
-};
-
-const noncesTrait: BaseImplementedTrait = {
-  name: 'Nonces',
-  storage: {
-    name: 'nonces',
+    name: NONCES_STORAGE_NAME,
     type: 'Nonces',
   },
   modulePath: 'openzeppelin_stylus::utils::nonces',
+  functions: Object.values(functions.nonces),
 };
 
-// const erc20MetadataTrait: BaseImplementedTrait = {
-//   name: 'Erc20Metadata',
-//   storage: {
-//     name: 'metadata',
-//     type: 'Erc20Metadata',
-//   }
-//   modulePath: 'openzeppelin_stylus::token::erc20::extensions',
-// }
+const permitTrait: StoredContractTrait = {
+  name: 'IErc20Permit',
+  associatedError: true,
+  errors: {
+    list: [
+      { variant: 'ExpiredSignature', value: { module: 'permit', error: 'ERC2612ExpiredSignature' } },
+      { variant: 'InvalidSigner', value: { module: 'permit', error: 'ERC2612InvalidSigner' } },
+      { variant: 'InvalidSignature', value: { module: 'ecdsa', error: 'ECDSAInvalidSignature' } },
+      { variant: 'InvalidSignatureS', value: { module: 'ecdsa', error: 'ECDSAInvalidSignatureS' } },
+    ],
+    wraps: erc20Trait.name,
+  },
+  storage: {
+    name: PERMIT_STORAGE_NAME,
+    type: 'Erc20Permit',
+    genericType: 'Eip712',
+  },
+  modulePath: 'openzeppelin_stylus::token::erc20::extensions::permit',
+  requiredImports: [
+    { containerPath: 'stylus_sdk::alloy_primitives', name: 'B256' },
+    { containerPath: 'openzeppelin_stylus::utils::cryptography', name: 'ecdsa' },
+  ],
+  functions: Object.values(functions.permit),
+};
 
-const functions = defineFunctions({
-  // Token Functions
-  transfer: {
-    args: [getSelfArg(), { name: 'to', type: 'Address' }, { name: 'value', type: 'U256' }],
-    returns: 'Result<bool, Vec<u8>>',
-    code: [`self.${erc20Trait.storage.name}.transfer(to, value).map_err(|e| e.into())`],
-  },
-  transfer_from: {
-    args: [
-      getSelfArg(),
-      { name: 'from', type: 'Address' },
-      { name: 'to', type: 'Address' },
-      { name: 'value', type: 'U256' },
-    ],
-    returns: 'Result<bool, Vec<u8>>',
-    code: [`self.${erc20Trait.storage.name}.transfer_from(from, to, value).map_err(|e| e.into())`],
-  },
+const burnableTrait: ContractTrait = {
+  name: 'IErc20Burnable',
+  associatedError: true,
+  modulePath: 'openzeppelin_stylus::token::erc20::extensions::burnable',
+  functions: Object.values(functions.burnable),
+};
 
-  // Extensions
-  burn: {
-    args: [getSelfArg(), { name: 'value', type: 'U256' }],
-    returns: 'Result<(), Vec<u8>>',
-    code: [`self.${erc20Trait.storage.name}.burn(value).map_err(|e| e.into())`],
-  },
-  burn_from: {
-    args: [getSelfArg(), { name: 'account', type: 'Address' }, { name: 'value', type: 'U256' }],
-    returns: 'Result<(), Vec<u8>>',
-    code: [`self.${erc20Trait.storage.name}.burn_from(account, value).map_err(|e| e.into())`],
-  },
-
-  permit: {
-    attribute: 'allow(clippy::too_many_arguments)',
-    args: [
-      getSelfArg(),
-      { name: 'owner', type: 'Address' },
-      { name: 'spender', type: 'Address' },
-      { name: 'value', type: 'U256' },
-      { name: 'deadline', type: 'U256' },
-      { name: 'v', type: 'u8' },
-      { name: 'r', type: 'B256' },
-      { name: 's', type: 'B256' },
+const flashMintTrait: StoredContractTrait = {
+  name: 'IErc3156FlashLender',
+  associatedError: true,
+  errors: {
+    list: [
+      { variant: 'UnsupportedToken', value: { module: 'flash_mint', error: 'ERC3156UnsupportedToken' } },
+      { variant: 'ExceededMaxLoan', value: { module: 'flash_mint', error: 'ERC3156ExceededMaxLoan' } },
+      { variant: 'ERC3156InvalidReceiver', value: { module: 'flash_mint', error: 'ERC3156InvalidReceiver' } },
     ],
-    returns: 'Result<(), Vec<u8>>',
-    code: [
-      `self.${erc20PermitTrait.storage.name}.permit(owner, spender, value, deadline, v, r, s, &mut self.${erc20Trait.storage.name}, &mut self.${noncesTrait.storage.name}).map_err(|e| e.into())`,
-    ],
+    wraps: erc20Trait.name,
   },
-
-  max_flash_loan: {
-    args: [getSelfArg('immutable'), { name: 'token', type: 'Address' }],
-    returns: 'U256',
-    code: [`self.${flashMintTrait.storage.name}.max_flash_loan(token, &self.${erc20Trait.storage.name})`],
+  storage: {
+    name: FLASH_MINT_STORAGE_NAME,
+    type: 'Erc20FlashMint',
   },
-  flash_fee: {
-    args: [getSelfArg('immutable'), { name: 'token', type: 'Address' }, { name: 'value', type: 'U256' }],
-    returns: 'Result<U256, Vec<u8>>',
-    code: [`self.${flashMintTrait.storage.name}.flash_fee(token, value).map_err(|e| e.into())`],
-  },
-  flash_loan: {
-    args: [
-      getSelfArg(),
-      { name: 'receiver', type: 'Address' },
-      { name: 'token', type: 'Address' },
-      { name: 'value', type: 'U256' },
-      { name: 'data', type: 'Bytes' },
-    ],
-    returns: 'Result<bool, Vec<u8>>',
-    code: [
-      `self.${flashMintTrait.storage.name}.flash_loan(receiver, token, value, data, &mut self.${erc20Trait.storage.name}).map_err(|e| e.into())`,
-    ],
-  },
-});
+  modulePath: 'openzeppelin_stylus::token::erc20::extensions::flash_mint',
+  requiredImports: [{ containerPath: 'stylus_sdk::abi', name: 'Bytes' }],
+  functions: Object.values(functions.flash_mint),
+};